Remove singleton SensorTag
REQ sockets are created in every function that needs to
access the SensorTag. Since the state is maintined in the BT GATT
daemon, we don't need to use singleton anymore. However, when the
page is refreshed, we get new SensorTag objects, and the threads
are starting up again. We need a way to make sure that the same
Thread is not started with the same arguments.