Global dict for storing threads
Each "thread" is only launched once. When you call it again
the global dict realizes that this is already running, and
doesn't run it again. However, this doesn't fix the GUI issue, since
the thread is connected to the previous widget, not the one that has
just been displayed. It therefore updates the old widget, and the
new one is not updated.